Removed legact *diff modules
The following modules were legacy, and removed:

* activitydiff
* appdiff
* providerdiff
* receiverdiff
* servicediff

This also means the binary appdiff.py is removed.  This will
result in a brief lack of ability to "diff" applications, but
in the meantime you can used the 'exposed' sub-command of the
`sysappdb` module.  Eventually `sysappdb` will support a 'diff'
sub-command that lets you compare an application's manifest.